move contexts to their own folder
