+++ frontend/interactEM/src/components/composerpipelineflow.tsx
@@ -30,7 +30,7 @@ import {
   pipelinesReadPipelineQueryKey,
   pipelinesReadPipelinesQueryKey,
 } from "../client/generated/@tanstack/react-query.gen"
-import { useDnD } from "../hooks/useDnD"
+import { useDnD } from "../contexts/dnd"
 import useOperators from "../hooks/useOperators"
 import { usePipelineStore } from "../stores"
 import { NodeType, type OperatorNodeTypes } from "../types/nodes"
