+++ frontend/interactEM/src/components/nodes/parameterupdater.tsx
@@ -15,9 +15,9 @@ import { useReactFlow } from "@xyflow/react"
 import type React from "react"
 import { memo, useEffect, useState } from "react"
 import type { OperatorParameter } from "../../client"
+import { usePipelineContext } from "../../contexts/pipeline"
 import { useParameterUpdate } from "../../hooks/useParameterUpdate"
 import { useParameterValue } from "../../hooks/useParameterValue"
-import { usePipelineContext } from "../../hooks/usePipelineContext"
 import { useEditModeState } from "../../stores/edit"
 import type { OperatorNodeType } from "../../types/nodes"
 import ParameterInfoTooltip from "./parameterinfotooltip"
