+++ frontend/interactEM/src/hooks/useBucket.ts
@@ -1,6 +1,6 @@
 import type { KV } from "@nats-io/kv"
 import { useEffect, useRef, useState } from "react"
-import { useNats } from "../nats/NatsContext"
+import { useNats } from "../contexts/nats"
 
 export const useBucket = (bucketName: string): KV | null => {
   const { keyValueManager } = useNats()
