+++ frontend/interactEM/src/hooks/useParameterUpdate.ts
@@ -5,7 +5,7 @@ import {
 } from "@tanstack/react-query"
 import type { OperatorParameter } from "../client"
 import { PARAMETERS_QUERYKEY, PARAMETERS_STREAM } from "../constants/nats"
-import { useNats } from "../nats/NatsContext"
+import { useNats } from "../contexts/nats"
 
 export const useParameterUpdate = (
   operatorID: string,
