+++ frontend/interactEM/src/pages/composerpage.tsx
@@ -6,10 +6,10 @@ import AgentsAccordion from "../components/agents/accordion"
 import ComposerPipelineFlow from "../components/composerpipelineflow"
 import { OperatorMenu } from "../components/operatormenu"
 import { PipelineHud } from "../components/pipelines/hud"
+import { usePipelineContext } from "../contexts/pipeline"
 import { useActivePipeline } from "../hooks/useActivePipeline"
 import { useAgents } from "../hooks/useAgents"
 import useOperators from "../hooks/useOperators"
-import { usePipelineContext } from "../hooks/usePipelineContext"
 import { usePipelineStore } from "../stores"
 import { useEditModeState } from "../stores/edit"
