+++ frontend/interactEM/src/pages/interactem.tsx
@@ -2,15 +2,15 @@ import { ReactFlowProvider } from "@xyflow/react"
 import ExternalAuthProvider from "../auth/externalprovider"
 import InternalAuthProvider from "../auth/internalprovider"
 import config from "../config"
-import { DnDProvider } from "../hooks/useDnD"
-import { NatsProvider } from "../nats/NatsContext"
+import { DnDProvider } from "../contexts/dnd"
+import { NatsProvider } from "../contexts/nats"
 
 import { QueryClientProvider } from "@tanstack/react-query"
 import "@xyflow/react/dist/style.css"
 import { Flip, ToastContainer } from "react-toastify"
 import { interactemQueryClient } from "../auth/api"
 import NotificationsToast from "../components/notificationstoast"
-import { PipelineProvider } from "../hooks/usePipelineContext"
+import { PipelineProvider } from "../contexts/pipeline"
 import "../index.css"
 import ComposerPage from "./composerpage"
